+++ bin/typedef/liff-update-request-data.js
@@ -1,0 +1,11 @@
+/**
+ * @typedef LIFFUpdateRequestData
+ * @property {object} view
+ * @property {'compact'|'tall'|'full'} view.type
+ * @property {string} view.url
+ * @property {string} description
+ * @property {object} features
+ * @property {boolean} features.ble
+ */
+"use strict";
+//# sourceMappingURL=liff-update-request-data.js.map